Move parser and loader to top level
