+++ packages/unmock-core/src/__tests__/generator.test.ts
@@ -1,6 +1,6 @@
-import fs from "fs";
 import path from "path";
-import { IServiceDefLoader, responseCreatorFactory } from "..";
+import { responseCreatorFactory } from "..";
+import { FsServiceDefLoader } from "../fs-service-def-loader";
 
 const mockOptions = {
   flaky: () => false,
@@ -9,29 +9,9 @@ const mockOptions = {
   useInProduction: () => false,
 };
 
-const serviceDefLoader: IServiceDefLoader = {
-  load: () => Promise.all(serviceDefLoader.loadSync()),
-  loadSync: () => {
-    const servicesDirectory: string = path.join(__dirname, "__unmock__");
-    const serviceDirectories = fs
-      .readdirSync(servicesDirectory)
-      .map((f: string) => path.join(servicesDirectory, f))
-      .filter((f: string) => fs.statSync(f).isDirectory());
-
-    return serviceDirectories.map((dir: string) => ({
-      absolutePath: dir,
-      directoryName: path.basename(dir),
-      serviceFiles: fs
-        .readdirSync(dir)
-        .map((fileName: string) => path.join(dir, fileName))
-        .filter((fileName: string) => fs.statSync(fileName).isFile())
-        .map((f: string) => ({
-          basename: path.basename(f),
-          contents: fs.readFileSync(f).toString("utf-8"),
-        })),
-    }));
-  },
-};
+const serviceDefLoader = new FsServiceDefLoader({
+  unmockDirectories: [path.join(__dirname, "__unmock__")],
+});
 
 describe("Tests generator", () => {
   it("loads all paths in __unmock__", () => {
