+++ packages/unmock-core/src/parser.ts
@@ -2,8 +2,8 @@ import debug from "debug";
 import { isLeft } from "fp-ts/lib/Either";
 import jsYaml from "js-yaml";
 import loas3 from "loas3";
-import { IServiceDef, IServiceDefFile } from "../interfaces";
-import { ServiceCore } from "./serviceCore";
+import { IServiceDef, IServiceDefFile } from "./interfaces";
+import { ServiceCore } from "./service/serviceCore";
 
 const debugLog = debug("unmock:service-parser");
