+++ packages/unmock-core/src/service/index.ts
@@ -1,3 +1,2 @@
-export { ServiceParser } from "./parser";
 export { Service } from "./service";
 export { IService, ServiceStoreType } from "./interfaces";
