--- packages/unmock-node/package.json
@@ -1,38 +1,0 @@
-{
-  "version": "0.2.0",
-  "name": "unmock-node",
-  "displayName": "unmock-node",
-  "main": "dist/index.js",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/unmock/unmock-js.git"
-  },
-  "homepage": "https://github.com/unmock/unmock-js",
-  "bugs": {
-    "url": "https://github.com/unmock/unmock-js/issues"
-  },
-  "dependencies": {
-    "app-root-path": "^2.2.1",
-    "debug": "^4.1.1",
-    "expect": "^24.7.1",
-    "ini": "^1.3.5",
-    "js-yaml": "^3.13.1",
-    "lodash": "^4.17.14",
-    "mitm": "^1.7.0",
-    "mkdirp": "^0.5.1",
-    "readable-stream": "^3.4.0",
-    "shimmer": "^1.2.1",
-    "unmock-core": "file:../unmock-core",
-    "uuid": "^3.3.2",
-    "winston": "^3.2.1"
-  },
-  "devDependencies": {
-    "@types/app-root-path": "^1.2.4",
-    "@types/debug": "^4.1.4",
-    "@types/readable-stream": "^2.3.2",
-    "@types/shimmer": "^1.0.1",
-    "@types/uuid": "^3.4.5",
-    "axios": "^0.19.0"
-  }
-}
